requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
azure-storage-blob
python-multipart
pandas
openpyxl
